support/scripts/nvd_api_v2.py: new helper class
The current NVD data feeds used for CVE and CPE checking will be retired
by 2023-12-05 [0]. Both have to be switched to the new v2 API. Since
fetching data from both sources workes the same, a common base class is
used to handle the API interaction.
With the new API JSON pages are downloaded, meaning that we have to come
up with a storage solution ourselves. Therefore nvd_api_v2.py manages a
generic set of methods to initialize and update a sqlite database.

[0]: https://nvd.nist.gov/General/News/change-timeline

Signed-off-by: Daniel Lang <dalang@gmx.at>
---
v3 -> v4:
- use f-strings
- add meta table instead of tmp file
- drop error handling